[cfid-72] Ensure that caller knows user password as well
The AutologinController accepts an optional AuthenticationManager
that it must use to verify the request it was given.  Any
AuthenticationManager will do (e.g. if password is not known
but the client has some other way to prove it knows the user)
but by default we use the existing RemoteUaaAuthenticationManager
which does a form login at the UAA.

[Fixes #35907911] [cfid-72] Autologin for trusted SSO clients that know the user