Move the selenium tests to a separate module
We found the maven tomcat plugin to be unreliable in starting up the integration test server.  This was causing mvn install to fail for the root project.  We decided to move the selenium tests out to a separate module, so that they can fail if explicitly run when the integration server hasn't been started.  This way mvn install succeeds, but we can run the integration tests on CI.